+++ scripts/macros/classFeatures/paladin/auraOfCourage.js
@@ -1,18 +1,16 @@
 import {chris} from '../../../helperFunctions.js';
-import {effectAura} from '../../../movement.js';
-async function effectCreation(token, actor, effect, origin) {
-    let castLevel = actor.system.abilities.cha.mod;
-    let sourceActorUuid = actor.uuid;
-    let range = 10;
-    let paladinLevels = actor.classes.paladin?.system?.levels;
-    if (!paladinLevels) return;
-    if (paladinLevels >= 18) range = 30;
-    let targetDisposition = 'ally';
-    let conscious = true;
+export async function auraOfCourage(token, selectedAura) {
+    let originToken = await fromUuid(selectedAura.tokenUuid);
+    let originActor = originToken.actor;
+    if (!originActor) return;
+    let auraEffect = chris.findEffect(originActor, 'Aura of Courage - Aura');
+    if (!auraEffect) return;
+    let originItem = await fromUuid(auraEffect.origin);
+    if (!originItem) return;
     let effectData = {
         'label': 'Aura of Courage',
-        'icon': origin.img,
-        'origin': origin.uuid,
+        'icon': originItem.img,
+        'origin': originItem.uuid,
         'duration': {
             'seconds': 604800
         },
@@ -25,22 +23,10 @@ async function effectCreation(token, actor, effect, origin) {
             }
         ]
     }
-    await effectAura.add('auraOfCourage', castLevel, 10, sourceActorUuid, range, targetDisposition, conscious, effectData, effect.uuid);
-}
-async function moved(token, castLevel, spellDC, effectData, tokenID) {
     let effect = chris.findEffect(token.actor, effectData.label);
     if (effect?.origin === effectData.origin) return;
     if (effect) chris.removeEffect(effect);
     await chris.createEffect(token.actor, effectData);
     let frightenedEffect = chris.findEffect(token.actor, 'Frightened');
     if (frightenedEffect) chris.removeEffect(frightenedEffect);
-}
-async function effectEnd(token ,effect) {
-    await effectAura.refresh(effect.uuid);
-    await effectAura.remove('auraOfCourage', token.actor.uuid);
-}
-export let auraOfCourage = {
-    'start': effectCreation,
-    'moved': moved,
-    'end': effectEnd
 }